Adding VIP cycle methods
------------------Exercise 2 - The VIP Cycle

The essence of Clean Swift is the VIP cycle. In this exercise, you’ll set up the VIP cycle for the ​*List Gists*​ scene.

Add a new `listGistsOnLoad()` method in your view controller and invoke it in `viewDidLoad(). Name the other methods in the VIP cycle as follows:

- `listGists()`
- `presentListGists()`
- `displayListGists()`

You are free to name them whatever you want in your own app. But I don’t want you to get hung up on naming while doing the exercise.